Copy configuration files for both normal and failover standby.
On the normal startup, postgresql_conf and pg_hba_conf options are
used. For the failover, it switches to the _failover counterparts.
In both cases, cmd_standby only attempts to copy files that exist
and readable.

Additionally, when no_copy_conf is set, add an option to use
pg_hba_conf file.